requirements.txt
PyYAML==6.0
tqdm==4.64.1
pandas==1.5.3
requests==2.28.2
